Add normalization for non-lemmatized text.

+++ src/Rank/ranking/util/json_lines.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import os
 
 DEFAULT_ENCODING = 'utf-8'
 
@@ -8,6 +9,7 @@ def read_dataset(source) -> pd.DataFrame:
     return df
 
 def write_dataset(df: pd.DataFrame, foutput):
+    if os.path.exists(foutput): raise ValueError('Output file already exists.') 
     to_jsonl(df.reset_index(), foutput)
 
 def read_jsonl(source, encoding=DEFAULT_ENCODING) -> pd.DataFrame:
@@ -15,6 +17,6 @@ def read_jsonl(source, encoding=DEFAULT_ENCODING) -> pd.DataFrame:
         df = pd.read_json(fin, lines=True)
     return df
 
-
 def to_jsonl(df: pd.DataFrame, foutput):
+    if os.path.exists(foutput): raise ValueError('Output file already exists.') 
     df.to_json(foutput, orient='records', lines=True, force_ascii=False)
